# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Accès Direct Aide API",
    "version": "1.0.0",
    "description": "Documentation complète de l'API de la plateforme Accès Direct Aide — aide sociale, démarches, structures, rendez-vous professionnels et administration.",
    "contact": {
      "name": "Équipe AccesDirectAide",
      "url": "https://www.accesdirectaide.fr"
    }
  },
  "servers": [
    {
      "url": "https://www.accesdirectaide.fr/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5173/api",
      "description": "Développement local (Vite proxy)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Développement local (API directe)"
    }
  ],
...
  "paths": {
...
    "/monitor/ingestion-freshness": {
      "get": {
        "tags": ["Monitor"],
        "summary": "Fraîcheur de l'ingestion de données",
        "responses": {
          "200": { "description": "Données fraîches" },
          "503": { "description": "Données obsolètes" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenu via /auth/login ou /pro/auth/login"
      },
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Token d'administration"
      },
      "CronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Bearer token cron (CRON_SECRET)"
      }
    }
  }
}